+++ Tests/SentryTests/TestClient.swift
@@ -11,9 +11,19 @@ class TestClient: Client {
 
     // Without this override we get a fatal error: use of unimplemented initializer
     // see https://stackoverflow.com/questions/28187261/ios-swift-fatal-error-use-of-unimplemented-initializer-init
-    override init(options: Options, transportAdapter: SentryTransportAdapter, fileManager: SentryFileManager, threadInspector: SentryThreadInspector, random: SentryRandomProtocol, crashWrapper: SentryCrashWrapper, permissionsObserver: SentryPermissionsObserver) {
+    override init(options: Options, transportAdapter: SentryTransportAdapter, fileManager: SentryFileManager, threadInspector: SentryThreadInspector, random: SentryRandomProtocol, crashWrapper: SentryCrashWrapper, permissionsObserver: SentryPermissionsObserver, locale: Locale, timezone: TimeZone) {
         sentryFileManager = try! SentryFileManager(options: options, andCurrentDateProvider: TestCurrentDateProvider())
-        super.init(options: options, transportAdapter: transportAdapter, fileManager: fileManager, threadInspector: threadInspector, random: random, crashWrapper: crashWrapper, permissionsObserver: permissionsObserver)
+        super.init(
+            options: options,
+            transportAdapter: transportAdapter,
+            fileManager: fileManager,
+            threadInspector: threadInspector,
+            random: random,
+            crashWrapper: crashWrapper,
+            permissionsObserver: permissionsObserver,
+            locale: locale,
+            timezone: timezone
+        )
     }
 
     override func fileManager() -> SentryFileManager {